feat: rename useAllowClear to getAllowClear
